Improved timing of custom display message handling for old-gen MT-32
If a program change or an error message is displaying, and a custom
display message is received, it is now bufferred, as the real hardware
does (so that it may be shown later with a special SysEx message).
Additionally, the display mode is reset at this point in case the timer
has already expired but the emulated display state hasn't been updated yet.