Display is now updated on any change of the master volume
In fact, real hardware does not show the updated master volume, if set via
SysEx, until the display is updated for some other reason. However, we have
no special API to set the master volume directly so far, yet this quirk is
unlikely desired to be emulated correctly.

Also, the display reset is now committed even when in the main mode,
same as real devices do.